Draw terrain on grids (mini elevation banding)
Grids are added. One banding texture is used to detect elevations. The
texture is drawn on grids. R and G channels of the texture represents
the elevation and the B channel is for light. So any other light
operation is not used. An elevation treshold is selected and colors of
the terrain are calculated accordingly.
Note: Model cannot be drawn with these shaders. A new shader is
necessary because objectPosition(y axis) is changed in vertex shader
now.