Move model on the terrain
Terrain height(y pos) is calculated in model shader as well. This height
value is added to object's y position in world space. However this value
is affected by scale of the model. In order to fix this issue, model
size is multiplied by scale size of the model and the result is also
added to y position.