Use framebuffer to see the scene from different perspectives
Framebuffer is used to see the scene from top of the model. A mini quad
is drawn on screen space for this and also a new shader is added.
Framebuffer is generated in initialization and a texture is attached to
it as color attachment. Then, the framabuffer is binded in main loop and
draw operations that we want to see the scene from top of the model, are
implemented. The framebuffer is unbinded and fbo texture is drawn on screen
space quad.
In brief, a mini screen is added to the scene and it always shows the
model from it's top. The view matrix is changed for this mini screen.

Cloud's finalheight is multiplied by -1 to see the elevations under
clouds.