Handle mouse button click event
When buttons are in the same position but in different states (for
example credits button in main menu and resume button is pause screen)
click event runs for both of them. Because it is only depend on the
position. In order to fix this issue, HandleMouseInput is added to
Button class. Therefore, click event is checked for each button
separetely now.
Note: There is a bug in click event. When you click outside the button
and release inside the button, click event runs. However, it should not
be a valid operation(click).
Since my main goal is to improve my opengl skills, this is not an
important issue. It can be fixed later.