Fix camera movement bug and add texture parser
New methods are added to update camera position in every mouse callback.
Also, a texture parser function is added into texture class. This class
will be used in order to read a texture and use it as the game map. For
ex. red pixels means walls in our game. According to map texture the
scene and models will be created.