Create game map
A 3-channel map texture is created using GIMP. Each unique pixel color corresponds to an entity. SetMapTexture() function is added to read map texture's pixel colors. In addition, translate values are calculated according to pixel coords. Left-bottom is 0.0 point in texture, however, in my game world 0,0 is the central point of the map. Therefore, translate values are mapped according to this transformation. Translate values are pushed to the related entity vector. In gameplaywindow, cube models are drawn to these entity translate coords.

With this method, the main scene of the game can be created independently of the map texture and the number of entitites.

Map is drawn to a quad. This quad vertices are set in mapShader, not in CPU.