Set camera like main character's eye
Camera movement is controlled by W,A,S,D keys.  Acceleration is added to velocity calculation. At the begining, velocity is 0. When the player starts to move, velocity increments and it stops in a constant speed value. Thanks to this new velocity calculation method, the player will feel like he/she is doing the real-world running action.