Implement normal map to walls
* XYZ axis of the scene is drawn in screen space
* Brickwall texture and it's normal map is added.
* Point light is applied properly.
* normal map is used as vertex normals.

It is aimed to see brickwall texture just like the same as the real world walls. However, the normals are different for each surface of the model. With this commit, the normal mapping can be seen inside of the models. This issue will be fixed later by using assimp and tangent space.