Add Assimp as model loader
The old model loader is removed and mesh class is changed to be compatible with the assimp lib. A new Model shader is added to draw assimp models and sponza, backpack models are loaded. The directional light function is added to model shader.
In addition, README is added.

Note: When build folder is deleted and new one is created, assimp-vc142-mtd.lib should be added to linker->input->additional dependencies in VS ide. Also, assimp-vc142-mtd.dll file should be copied to build/Source/Debug folder. (same path with the application exe.)