Shadow Mapping
Shadow of backpack model is drawn. First, depth values are drawn to the FBO depthMap texture. It is shown in the left top corner of the window in screen space. Then the scene is drawn two times. The first time, the scene is drawn from the light's view and the second time, scene is drawn from the camera's view. Light direction can be changed from imgui. Also, default ortho of the light view is drawn with pink lines.