Cubemap
Cubemap is implemented using skybox textures and cubeMapShader. Model matrix is not used in cubemapShader because the cubemap is kind of a scene and it should not be affected from any of the rotate/translate/scale operations. The other models(like f16 in this commit) can be move in the scene but the cubemap should be seen as fixed.